refactor: main functionality using Path
This commit introduces a major refactor in the main file of Whaler. That
is instead of using multiple function, the new scanning returns exactly
what it required, which is a list of directories (string[]).

More documentation has been added to multiple functions clarifying
the goal of thy.

Better error handling in general. If something fails, now whaler is more
explicit about the errors.

The State is now properly populated on setup allowing use of Whaler without
requiring the use of `whaler` first.